Added the settings configuration file
settings.json will be kept in the installer directory, instead of rayshud. This will correct a design issue where user settings would be overwritten if they were made prior to installing rayshud. This way also removes the need for backups and makes it easier to download a fresh copy of the settings file from GitHub should it be removed from the installer director.